Perform casts or variable type conversions in code manipulating file
positions in order to adapt for -Wconversion.

This closes #46.

+++ core/Lucy/Index/LexIndex.c
@@ -83,7 +83,7 @@ LexIndex_init(LexIndex *self, Schema *schema, Folder *folder,
     }
     ivars->index_interval = Arch_Index_Interval(arch);
     ivars->skip_interval  = Arch_Skip_Interval(arch);
-    ivars->size    = (int32_t)(InStream_Length(ivars->ixix_in) / sizeof(int64_t));
+    ivars->size    = (int32_t)(InStream_Length(ivars->ixix_in) / (int32_t)sizeof(int64_t));
     ivars->offsets = (const int64_t*)InStream_Buf(ivars->ixix_in,
             (size_t)InStream_Length(ivars->ixix_in));
